+++ Changelog.md
@@ -1,5 +1,11 @@
 ## Ignition Gazebo 1.x
 
+### Ignition Gazebo 2.X.X
+
+1. Update Magnetometer component to use sdf::Sensor object instead of an
+   sdf::ElementPtr.
+    * [Pull Request 272](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/272)
+
 ### Ignition Gazebo 1.X.X
 
 1. More ign-msgs <-> SDF conversions: Inertial, Geometry, Material